src/microservices/consumptions/requirements.txt
scikit-learn~=1.1.3
numpy~=1.23.4
pandas~=1.5.1
grpcio-tools
grpcio
jproperties
cassandra-driver